More rustup.
This switches to using the new `std::convert` traits (mostly instread of
`std::borrow::IntoCow`). It seems like there are a few missing impls, so
you can't give the `Docopt::args` method an iterator of `String`.
You can probably work-around it for now by giving an iterator of `&str`.
(So this might incur an extra copy, but it's only temporary, I hope.)